被攻击增加读条，减少引导条

+++ Packages/cn.etetet.proto/CodeMode/Model/Client/WOWOuter_C_4100.cs
@@ -1069,7 +1069,7 @@ namespace ET
         public long BuffId { get; set; }
 
         [MemoryPackOrder(4)]
-        public long TickTime { get; set; }
+        public int TickTime { get; set; }
 
         [MemoryPackOrder(5)]
         public long ExpireTime { get; set; }
